Loadout, integer HP/durability, archetype-driven NPC gear
Replace Inventory with Loadout (typed slots) for NPCs and Stash for
bulk storage. Switch Health and item durability to integer budgets
(default 100 HP, Option<u32> for indestructible items). Add Rank enum.
Resistances::resolve_hit returns (dealt, absorbed) HP via a multiplicative
penetration ratio with a min-1 floor. Authors weapons/consumables JSON
and an archetype-per-faction catalog so each NPC spawns with rank-
appropriate gear rolled at NPC creation time.

+++ crates/cordon-core/src/item/data/ammo.rs
@@ -14,13 +14,12 @@ use crate::primitive::Id;
 pub struct AmmoData {
     /// Caliber ID this ammo belongs to (e.g., `"9x18mm"`).
     pub caliber: Id<Caliber>,
-    /// Base damage per round.
-    pub damage: f32,
+    /// Base damage per round in HP.
+    pub damage: u32,
     /// Armor penetration value. Compared against armor's protection
-    /// value — if penetration exceeds armor, full damage goes through.
-    /// If lower, damage is reduced proportionally.
+    /// value via [`Resistances::resolve_hit`](crate::primitive::Resistances::resolve_hit).
     /// Typical range: 5 (pistol) to 40 (AP rifle).
     pub penetration: u32,
-    /// Number of rounds per box.
+    /// Number of rounds per fresh box.
     pub quantity: u32,
 }
